Fix types for all fields

+++ fetch_example_json.py
@@ -4,6 +4,7 @@
 import os
 import shutil
 
+from lowerpines.endpoints.block import Block
 from lowerpines.endpoints.bot import Bot
 
 from lowerpines.endpoints.group import Group
@@ -25,6 +26,8 @@ except ImportError:
 
 TEST_DATA_FOLDER = "test_data"
 
+allow_non_deterministic = True
+
 if __name__ == "__main__":
     gmi = get_gmi(TEST_ACCESS_TOKEN)
 
@@ -32,23 +35,24 @@ if __name__ == "__main__":
     if os.path.exists(TEST_DATA_FOLDER):
         shutil.rmtree(TEST_DATA_FOLDER)
 
-    TEST_GROUP_NAME = "TestGroup"
-    try:
-        test_group = gmi.groups.get(name=TEST_GROUP_NAME)
-    except NoneFoundException:
-        test_group = Group(gmi, name=TEST_GROUP_NAME)
-        test_group.save()
+    if allow_non_deterministic:
+        TEST_GROUP_NAME = "TestGroup"
+        try:
+            test_group = gmi.groups.get(name=TEST_GROUP_NAME)
+        except NoneFoundException:
+            test_group = Group(gmi, name=TEST_GROUP_NAME)
+            test_group.save()
 
-    TEST_BOT_NAME = "TestBot"
-    try:
-        test_bot = test_group.bots.get(name=TEST_BOT_NAME)
-    except NoneFoundException:
-        test_bot = Bot(gmi, group_id=test_group.group_id, name=TEST_BOT_NAME)
-        test_bot.save()
+        TEST_BOT_NAME = "TestBot"
+        try:
+            test_bot = test_group.bots.get(name=TEST_BOT_NAME)
+        except NoneFoundException:
+            test_bot = Bot(gmi, group_id=test_group.group_id, name=TEST_BOT_NAME)
+            test_bot.save()
 
-    test_bot.post("BotMessage")
-    test_group.post("UserMessage")
-    test_group.messages.recent()
+        test_bot.post("BotMessage")
+        test_group.post("UserMessage")
+        test_group.messages.recent()
 
     gmi.refresh()
 
@@ -57,5 +61,12 @@ if __name__ == "__main__":
     gmi.bots.filter()
     gmi.chats.filter()
 
-    test_bot.delete()
-    test_group.delete()
+    my_user_id = gmi.user.get().user_id
+
+    Block(gmi).get_all(my_user_id)
+    Block.block(gmi, my_user_id, "6911718")
+    Block.block_exists(gmi, my_user_id, "6911718")
+
+    if allow_non_deterministic:
+        test_bot.delete()
+        test_group.delete()
